Revamp of main GUI.
Ubuntu's Unity window manager does not have a task bar, which leaves
the Plover task bar icon without a home. To overcome this, the Plover
task bar icon has been replaced with a minimal frame with the same
functionality. This change was tested on Ubuntu 11.10. This change is
likely a useful step toward Plover running on Mac OS X, though this
hasn't been tested.