Refactoring the ConfigDialog class to use different modalities.
The bind on the self._cancel button was provoking strange behaviours
when the ConfigurationDialog was modal. I removed it at no cost for the
actual fonctionnalities (clicking on cancel will still close the
dialog).

I added a EndModal call that provides a return value of wx.ID_SAVE for
the ShowModal call. Otherwise, the value returned to ShowModal() would
be wx.ID_CANCEL for both _save() and _cancel() methods.