Optionally grab keyboard on X11 using XGrabKeyboard, fixes #41
The upside: delete emulation is no longer necessary!

The downside: GTK's default textbox seems to hide the cursor when
someone else has grabbed the keyboard (for example, when using gedit),
so the editing experience gets worse for many situations.  As a
result, use of this feature is optional, and OFF by default.

Signed-off-by: Edward Z. Yang <ezyang@cs.stanford.edu>