Add case mode and space character with meta commands
Introduce a new breed of meta command, "MODE". Now Plover has persistent "modes" that affect the output of the text. The core of the functionality comes from controlling just:

- The text case
- The space character

For the text case, users can use "CAPS", "TITLE", "LOWER". This lets the text come out IN ALL CAPS, In Title Case, and, in. lower case.

For the space character, using "SET_STRING", users can effectively replace Plover's built in "space" character with any of their choosing. For example, youcouldsetthecharacter to nothing, or-to-the-dash, or-ow-to-ow-any-ow-arbitrary-ow-string.

Then, to disable the functionality, users can use "MODE:RESET_SPACE" to reset to the default space, "MODE:RESET_CASE" to reset to the default case, and "MODE:RESET" to clear the space character and the case.

Finally, Plover ships with some special mode names to help out programmers: "SNAKE" and "CAMEL". Snake just sets the space character to underscores,_like_this_if_you_will, and Camel sets the next character to lowercase (using "{>}"), sets the case to Title case, and sets the space character to none, thus resulting inThisOutputStyle.

Example dictionary for most of the strokes (reset space and case aren't useful to me personally):

```json
{
 "KA*PS": "{MODE:CAPS}",
 "T-LT": "{MODE:TITLE}",
 "HR*R": "{MODE:LOWER}",
 "STPH-BG": "{MODE:SNAKE}",
 "R-FT": "{MODE:RESET}",
 "KHRAO*ER": "{MODE:CLEAR}",
 "TK-RBS": "{MODE:SET_SPACE:-}",
 "K-BGS": "{MODE:CAMEL}",
 "TPH-FPS": "{MODE:SET_SPACE:}"
}
```