Reduce Tréal CPU usage
The Tréal file was set to use non-blocking reads from the hidapi. But that code was in a while loop, and so it was called infinitely, causing very high CPU usage. I've switched over to a blocking version of the call, which actually gives Plover a break in the while loop.

The effects have only been positive so far. I feel a little silly for not noticing this high CPU usage before in the months using Plover with the Tréal 😕