oslayer/xkeyboardcontrol: add Unicode support
This is done by dynamically changing the keymap when a character we want
to send is not already mapped. A rolling list of unused mappings is used
to reduce the risk of reusing a keycode+modifiers combo before the target
application has time to process each sequence of keymap change + key
press/release events. The keymap entries used by Plover are marked with
a special keysym, so if Plover is killed/restarted, those entries are
still considered unused, and consequently reusable.

Note that this also correct the Unicode character to keysym conversion
code, so non-latin1 characters present in the X11 keymap will be
correctly handled.