Switch to x64 builds on Mac
We were using 32-bit builds due to an old version of wxPython.
Additionally, PyObjc had an update that allowed us to use
"CGEventKeyboardSetUnicodeString" without a conversion through
PyObjc.

The application could still technically be built with 32-bit
support in "universal" fashion, but that value is super marginal,
because it's more complex to get setup and this would only affect
users on 32-bit OSX machines, which are very few. If they insist
on 32-bit, they can use the older builds.

The other reason to stick with a stripped application is the
filesize. A universal OSX application weighs in at about 90MB,
the 64/32-bit-only versions are a lighter 50MB (less in a .dmg).