Rename "NKRO Keyboard" and "Sidewinder" to "Keyboard"
In the code, we still refer to the Sidewinder even though it is no longer relevant for most users. As such, it's a little confusing when you are looking for the "keyboard" machine and you need to learn that it's called the Sidewinder internally. Also, it's a little misleading to say "NKRO Keyboard: connected", because users who don't have an NKRO keyboard will open Plover and see this message, and may be confused.

These issues affected me when I very first started using Plover (no NKRO keyboard, looking to find the plover/machine/keyboard).