Fix currency retroformatting exception
When trying to currency format after an output that wasn't a number, there
was an exception and Plover's output would cease completely.

To work around this is just to define the case where the user's previous
word is not a number, at which point the command should do nothing.
Explicitly added tests for the 0 cases as well, because they are edge.

Also refactor the code into a function instead of having it in two places.