oslayer/winkeyboardcontrol: improve keyboard capture
Even if the keyboard capture subprocess is set as 'daemon', there is no
guarantee it will be terminated if its parent process gets violently
killed; e.g. by a signal (as Python 'atexit' code will not be executed).

This can be observed by starting Plover from source using launch.bat,
enabling output, and then interrupting the batch file with Control+C:
the main Plover process is killed, but the capture process is still
alive (and keeps on suppressing the keyboard).

This is fixed by adding some code in the subprocess to monitor its
parent: if the parent dies, the subprocess will terminate itself
automatically.