osxkeyboardcontrol: Name both threads
These names show up on the console prefixed to any log messages emitted.

One of the threads was already named in standard OS X reverse-DNS style.
I backed that out in favor of more direct names since the Python thread
names aren't surfaced as system thread names,
so the names don't need to conform to the same style as expected
for OS threads.