Ignore only KeyPress when a modifier is set
Previously, both KeyPress and KeyRelease were ignored if there were any
modifiers. This runs into the issue of a key "sticking" in a perpetually
pressed state if the key is pressed without modifiers but released with
modifiers.

This wouldn't be an issue in steno, but it creates problems when
switching in and out of steno with {PLOVER:TOGGLE} and typing rapidly
with a conventional keyboard layout. (See the GitHub issue referenced
for further details.)

Fixes #597.