Implement arpeggiate_group config option
This is the first step of implementing #591. The arpeggiate_group option
is an array listed under [Keyboard] in plover.cfg, which contains a list
of keys that should trigger arpeggiation.

Currently, the option works when an arpeggiation key is the last key
pressed in a chord, but it does not work in the middle of a chord.

The option is not yet exposed in the configuration GUI.