Add support for "stitching" mode
To "stitch" is to intersperse a given character between each of the
letters you're stroking. For example, one might stitch a hyphen into a
word being spelled:

    Plover is spelled P-L-O-V-E-R.

Or they might put periods in between the letters of an initialism:

    The United States of America is abbreviated U.S.A.

The command can be used with a dictionary entry in the form
`{MODE:SET_STITCH:-}`.