Gemini PR sync problems
So the basic sync problem is there was no attempt to, well,
sync. To sync, we have to look for a 6-byte sequence starting
with 0x80.

Hypothesized cause of sync failures is getting partial reads.
Since the code always just read the next six bytes, it'd never
be in sync again.

So, we attempt to properly resync.

If we end up with a buffer which does not start with a byte
with 0x80 set, or which has another byte with 0x80 set, or
which is not six bytes, we look for the last byte with 0x80
set, and stash that plus any following bytes, and try to fill
out a 6-byte packet by adding on to that.

If we get a partial read, but more bytes are forthcoming,
we will stash the entire partial read in saved, then fill
it in, and everything works.

If we get a partial read, and at least one byte goes missing:
  * first run of loop: we get a partial read. we stash the
    whole thing.
  * next run of loop: we do *not* finish this read, but we
    get part of the next packet, which starts with an 0x80
    byte. this is invalid, so we stash that 0x80 plus any
    following bytes
  * next run of loop: we finish that packet and are back in
    sync.

So no matter what sequence of bytes we get, we will process
every valid 6-byte packet starting with an 0x80 byte. I
think.

Signed-off-by: seebs <seebs@seebs.net>