Tooltips: consistent punctuation and localizability
Make punctuation in tooltips consistent between files. If one contains a whole sentence or several, end it with a full stop (period).

Also, give the Tool class a TOOLTIP member variable and put tooltips there rather than in docstrings, since docstrings were meant as a type of comments shown development tools rather than the return value of a function, removing the docstrings that are no longer used as tooltips.

Remove a comment about using docstrings as tooltips, and add one that mentions that the TOOLTIP member variable is used for tooltips shown in the main window.