Add tests for lookup_from_all. Fix dictionary collection filtering.
Added tests for lookup_from_all. A bug was also found in the filtering of
key value pairs in the dictionary collection lookup-function that was not
picked up by the tests due to wrong assertion. If a value was fitterred out
from one dictionary, no other value were returned even if a value that is
not supposed to be filtered out appears in a different dictionary.

The filtering is now fixed.