Update README.md
Here are the steps required to get things setup on a recent Ubuntu 18.04 install. I haven't tried it on a completely fresh install, and it's possible that I missed some system dependencies. I'm also not totally sure if `python-dbus` is required, and I assumed that versions of python newer than 3.6 would work fine.