Show dictionary on add translation (#1084)
When adding new or checking existing stroke, definitions the dialog
box shows also the dictionary file name excluding the path of the
current mapping if it exists. It also lists other mappings if present
in other dictionaries.

Translation text style is changed to bold for better visibility.

Added tests for lookup_from_all. A bug was also found in the filtering of
key value pairs in the dictionary collection lookup-function that was not
picked up by the tests due to wrong assertion. If a value was filtered out
from one dictionary, no other value were returned even if a value that is
not supposed to be filtered out appears in a different dictionary.

The filtering is now fixed.

Co-authored-by: Ted Morin <morinted@gmail.com>