doc: Add docs for installing and building from source and using plugins
This is pretty much copied from the existing wiki, with some changes
regarding macOS 11 Big Sur, since the pre-built Plover binaries don't
currently work on this platform and will have to be built from source.

This grouping also allows me to split up the docs into a "quick start"
for people who just want to use Plover as-is, and an advanced usage
section involving things like plugins and non-English systems.