gui_qt/dictionaries: prevent certain operations on loading dictionaries
No editing or "saving as" until a dictionary is loaded.

Note: this does not strictly prevent trying to edit a loading
dictionary because the code does not filter the list of selected
dictionaries before attempting to do so. But this is use case will
result in the interface freezing until all selected dictionaries have
finished loading (and then the editor will show up).