Don't crash if _update_layout fails
On macOS and with some input sources (Bulgarian BDS in particular),
keyboardlayout.py doesn't properly parse the layout, throws an error,
and makes Plover crash:

    Traceback (most recent call last):
      File "/Users/skanev/code/opensource/plover/plover/scripts/main.py", line 147, in main
        code = gui.main(config, controller)
      File "/Users/skanev/code/opensource/plover/plover/gui_qt/main.py", line 125, in main
        app = Application(config, controller, use_qt_notifications)
      File "/Users/skanev/code/opensource/plover/plover/gui_qt/main.py", line 61, in __init__
        config, controller, KeyboardEmulation()
                            ~~~~~~~~~~~~~~~~~^^
      File "/Users/skanev/code/opensource/plover/plover/oslayer/osx/keyboardcontrol.py", line 412, in __init__
        self._layout = KeyboardLayout()
                       ~~~~~~~~~~~~~~^^
      File "/Users/skanev/code/opensource/plover/plover/oslayer/osx/keyboardlayout.py", line 130, in __init__
        self._update_layout()
        ~~~~~~~~~~~~~~~~~~~^^
      File "/Users/skanev/code/opensource/plover/plover/oslayer/osx/keyboardlayout.py", line 156, in _update_layout
        KeyboardLayout._get_layout()
        ~~~~~~~~~~~~~~~~~~~~~~~~~~^^
      File "/Users/skanev/code/opensource/plover/plover/oslayer/osx/keyboardlayout.py", line 268, in _get_layout
        parsed_layout = KeyboardLayout._parse_layout(layout_buffer, keyboard_type)
      File "/Users/skanev/code/opensource/plover/plover/oslayer/osx/keyboardlayout.py", line 423, in _parse_layout
        mod = modifier_masks[i]
              ~~~~~~~~~~~~~~^^^
    KeyError: 9

This happens both when Plover is started while in Bulgarian, and also
when switching from English to Bulgarian.

This change suppresses the crash when the input source is changed. That
way you can start Plover in English, and if you switch to Bulgarian, it
would not crash. It will output gibberish in Cyrillic, but that's OK, as
there is no Bulgarian steno yet anyway. Plover still crashes when
started in Bulgarian, but that can be worked around by starting it in
English.